# [Perf] Add support for int comparisons in StatefulExpression.EvalBool