Hi,

While adding the benchmarks to AlertNode (see #407 ), I found out the EvalPredicate is a tiny bottleneck, 
so I dug into this, and found out that StatefulExpression dosen't support int comparison so I added it and it improved the perf a bit ;) 

**StatefulExpression EvalBool benchmarks:**
```
benchmark                             old ns/op     new ns/op     delta
Benchmark_EvalBool_1ScopeItem-4       3069          3033          -1.17%
Benchmark_EvalBool_11_ScopeItem-4     3043          2708          -11.01%

benchmark                             old allocs     new allocs     delta
Benchmark_EvalBool_1ScopeItem-4       4              4              +0.00%
Benchmark_EvalBool_11_ScopeItem-4     4              4              +0.00%

benchmark                             old bytes     new bytes     delta
Benchmark_EvalBool_1ScopeItem-4       64            64            +0.00%
Benchmark_EvalBool_11_ScopeItem-4     64            64            +0.00%
```

**Alert benchmark:**
```
benchmark                            old ns/op       new ns/op       delta
Benchmark_T10_P500_AlertTask-4       145918779       143226340       -1.85%
Benchmark_T10_P50000_AlertTask-4     13739584526     13095020431     -4.69%
Benchmark_T1000_P500_AlertTask-4     15098593884     12948902710     -14.24%

benchmark                            old allocs     new allocs     delta
Benchmark_T10_P500_AlertTask-4       466191         466201         +0.00%
Benchmark_T10_P50000_AlertTask-4     47548950       47548763       -0.00%
Benchmark_T1000_P500_AlertTask-4     46101716       46098518       -0.01%

benchmark                            old bytes      new bytes      delta
Benchmark_T10_P500_AlertTask-4       32825881       32826965       +0.00%
Benchmark_T10_P50000_AlertTask-4     3344051504     3344020128     -0.00%
Benchmark_T1000_P500_AlertTask-4     3275554016     3274434464     -0.03%
```

_Those benchmarks are compared to the master (the edge buffer size is the default value)_