@nathanielc 2016-04-05
@yosiat I am a little confused:

Int comparisons worked before but were always casted to floats if one of the operands was a float. Is the PR to add new functionality or show that by not casting to float and doing direct integer comparisons things are faster?