@yosiat 2016-04-05
@nathanielc It started by reading the code of evalBinary and found out there is cast to float64 even when both sides are int64, so I changed this for clarity, I found this very confusing.
And then I checked out the performance, and as we can see it improved.

The initial purpose was to improve performance.