@nathanielc 2016-04-05
> The initial purpose was to improve performance.

Ok, makes sense.