@yosiat 2016-04-06
@nathanielc I think we should merge this PR, because we don't know when we will finish the "specialization" pr and this pr yields some nice perf improvement.