@yosiat 2016-04-07
@nathanielc I am closing this PR, since I will do this improvement on the typed stack approach (I will submit for this today/tomorrow)