@yosiat 2016-04-14
@nathanielc What is the status of this PR? I think we should merge it..