@nathanielc 2016-04-14
I think its ready, but first lets merge #458 since its now has tests for this change. Then a rebase to see we are green and this should be good to go.