@yosiat 2016-04-14
@nathanielc I squashed :+1: 

By the way, those are the results for this change on the new benchmarks that I am working on (https://github.com/yosiat/kapacitor/commit/50a94c906a93b984d2449e45646492acae3009d4):
```
name                                                                   old time/op    new time/op    delta
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                        1.85µs ± 0%    1.66µs ± 0%  -10.26%
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                          1.64µs ± 0%    1.63µs ± 0%   -0.43%
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                 2.00µs ± 0%    1.93µs ± 0%   -3.65%
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4            1.94µs ± 0%    2.02µs ± 0%   +4.34%
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4    2.12µs ± 0%    2.20µs ± 0%   +3.78%

name                                                                   old alloc/op   new alloc/op   delta
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                         80.0B ± 0%     80.0B ± 0%   +0.00%
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                           80.0B ± 0%     80.0B ± 0%   +0.00%
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                  64.0B ± 0%     64.0B ± 0%   +0.00%
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4             64.0B ± 0%     64.0B ± 0%   +0.00%
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4     64.0B ± 0%     64.0B ± 0%   +0.00%

name                                                                   old allocs/op  new allocs/op  delta
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                          5.00 ± 0%      5.00 ± 0%   +0.00%
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                            5.00 ± 0%      5.00 ± 0%   +0.00%
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                   4.00 ± 0%      4.00 ± 0%   +0.00%
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4              3.00 ± 0%      3.00 ± 0%   +0.00%
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4      4.00 ± 0%      4.00 ± 0%   +0.00%                                                                                                         
```
_I compared using benchstat instead of benchcmp_