@yosiat 2016-04-17
I am closing this pull request because I am going to submit another pull request that improve performance vastly.

After the another pull request will get merged we can decide if we want this pull request.